+++ packages/builder/support/logger.js
@@ -1,9 +1,7 @@
 const chalk = require('chalk')
 
-const logger = phrase => {
+module.exports = function logger(phrase) {
   const d = new Date()
   const timeStr = `[${d.toTimeString().split(' ')[0]}]`
   return `${chalk.gray(timeStr)} ${phrase || ''}`
 }
-
-module.exports = logger